Add dialog for allowing partial purchases of reserved LIFT
